create InteractionAwareDialogFragment
