+++ common/build.gradle
@@ -1,4 +1,6 @@
 apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
 
 android {
     compileSdkVersion 29
@@ -51,4 +53,7 @@ dependencies {
     androidTestImplementation('androidx.test.espresso:espresso-core:3.3.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
+
+    implementation "androidx.core:core-ktx:1.3.2"
+    //implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version
 }
